+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.19
 - [cpp] added new `cpp.clangdExecutable` and `cpp.clangdArgs` to customize language server start command.
 - [monaco] Fix document-saving that was taking too long.
+- [plug-in] added `tasks.onDidStartTask` Plug-in API
 
 ## v0.3.18
 - [core] added a preference to define how to handle application exit